requirements.txt
gtts
sox
deep_translator
google-cloud-texttospeech
tqdm
srt~=3.5.2
